+++ CHANGELOG.md
@@ -130,6 +130,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `jj new` now also accepts a `--allow-large-revsets` argument that behaves
   similarly to `jj rebase --allow-large-revsets`.
 
+* `jj new --insert-before` inserts the new commit between the target commit and
+  its parents.
+
 ### Fixed bugs
 
 * When sharing the working copy with a Git repo, we used to forget to export